requirements.txt
streamlit>=1.37.0
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
joblib>=1.3
Pillow>=10.0
matplotlib>=3.8
plotly>=5.15
fastapi>=0.110
uvicorn>=0.30
python-multipart>=0.0.9
# External HTTP requests
requests>=2.31
# MQTT client
paho-mqtt>=1.6.1
# TensorFlow removed (not available for Python 3.14). If needed, install separately on a supported Python version (<=3.12) or add to an optional requirements file.